separate testing suites and adjust workflows
