+++ .github/workflows/live-service-testing.yml
@@ -21,9 +21,12 @@ on:
         required: true
       DANDI_API_KEY:
         required: true
+      EMBER_API_KEY:
+        required: true
 
 env:
   DANDI_API_KEY: ${{ secrets.DANDI_API_KEY }}
+  EMBER_API_KEY: ${{ secrets.EMBER_API_KEY }}
 
 jobs:
   run:
@@ -61,6 +64,8 @@ jobs:
 
       - name: Run subset of tests that use DANDI live services
         run: pytest -rsx -n auto tests/remote_transfer_services/dandi_transfer_tools.py
+      - name: Run subset of tests that use EMBER live services
+        run: pytest -rsx -n auto tests/remote_transfer_services/ember_transfer_tools.py
       - name: Run subset of tests that use DANDI live services with YAML
         run: pytest -rsx -n auto tests/remote_transfer_services/yaml_dandi_transfer_tools.py
       - name: Run subset of tests that use Globus live services
