+++ tests/remote_transfer_services/dandi_transfer_tools.py
@@ -12,10 +12,12 @@ from neuroconv.tools.nwb_helpers import (
     make_nwbfile_from_metadata,
 )
 
-DANDI_API_KEY = os.getenv("DANDI_API_KEY")
 EMBER_API_KEY = os.getenv("EMBER_API_KEY")
+if EMBER_API_KEY is not None:
+    del os.environ["EMBER_API_KEY"]  # Will only remove within this test run
+
+DANDI_API_KEY = os.getenv("DANDI_API_KEY")
 HAVE_DANDI_KEY = DANDI_API_KEY is not None and DANDI_API_KEY != ""  # can be "" from external forks
-HAVE_EMBER_KEY = EMBER_API_KEY is not None and EMBER_API_KEY != ""  # can be "" from external forks
 
 
 @pytest.mark.skipif(
@@ -84,6 +86,10 @@ def test_automatic_dandi_upload_non_parallel_non_threaded(tmp_path):
     )
 
 
+@pytest.mark.skipif(
+    not HAVE_DANDI_KEY,
+    reason="You must set your DANDI_API_KEY to run this test!",
+)
 def test_staging_sandbox_conflict(tmp_path):
     """Test that providing both 'staging' and 'sandbox' parameters raises ValueError."""
 
@@ -122,22 +128,3 @@ def test_staging_backward_compatibility(tmp_path):
         # Check that deprecation warning was issued
         deprecation_warnings = [warning for warning in w if issubclass(warning.category, DeprecationWarning)]
         assert len(deprecation_warnings) == 1, f"Expected 1 deprecation warning, got {len(deprecation_warnings)}"
-
-
-@pytest.mark.skipif(
-    not HAVE_EMBER_KEY,
-    reason="You must set your DANDI_API_KEY to run this test!",
-)
-def test_automatic_ember_upload(tmp_path):
-    nwb_folder_path = tmp_path / "test_nwb"
-    nwb_folder_path.mkdir()
-    metadata = get_default_nwbfile_metadata()
-    metadata["NWBFile"].update(
-        session_start_time=datetime.now().astimezone(),
-        session_id=f"test-automatic-upload-{sys.platform}-{get_python_version().replace('.', '-')}",
-    )
-    metadata.update(Subject=dict(subject_id="foo", species="Mus musculus", age="P1D", sex="U"))
-    with NWBHDF5IO(path=nwb_folder_path / "test_nwb_1.nwb", mode="w") as io:
-        io.write(make_nwbfile_from_metadata(metadata=metadata))
-
-    automatic_dandi_upload(dandiset_id="000431", nwb_folder_path=nwb_folder_path, instance="ember")
